refactor(toolchain): isolate LLVM block map parser
